requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.2.10           # via -r requirements.in, django
attrs==19.3.0             # via pytest
django-environ==0.4.5     # via -r requirements.in
django-filter==2.3.0      # via -r requirements.in
django==3.1               # via -r requirements.in, django-filter, djangorestframework, djangorestframework-simplejwt
djangorestframework-simplejwt==4.4.0  # via -r requirements.in
djangorestframework==3.11.1  # via -r requirements.in, djangorestframework-simplejwt
gunicorn==20.0.4          # via -r requirements.in
iniconfig==1.0.1          # via pytest
more-itertools==8.4.0     # via pytest
packaging==20.4           # via pytest
pluggy==0.13.1            # via pytest
psycopg2-binary==2.8.5    # via -r requirements.in
py==1.9.0                 # via pytest
pyjwt==1.7.1              # via -r requirements.in, djangorestframework-simplejwt
pyparsing==2.4.7          # via packaging
pytest-django==3.9.0      # via -r requirements.in
pytest==6.0.1             # via pytest-django
pytz==2020.1              # via -r requirements.in, django
six==1.15.0               # via packaging
sqlparse==0.3.1           # via -r requirements.in, django
toml==0.10.1              # via pytest

# The following packages are considered to be unsafe in a requirements file:
# setuptools
